Fix assignment-related issues for student roles: - Resolved avatar retrieval issue for the class an assignment belongs to. - Fixed date and time display for assignments in student role. - Refactored navigation structure for student assignment section. - Display overdue and completed assignments sorted from the most recent date. - Corrected mix-up between lecturer and student assignment files.
