+++ app/student/(tabs)/assignments/(tabs)/overdueAssignments.tsx
@@ -4,47 +4,44 @@ import AssignmentCard from '@/components/AssignmentCard';
 import { Assignment, fetchAssignments } from '@/services/api-calls/assignments';
 import { router, useFocusEffect } from 'expo-router';
 import { useErrorContext } from '@/utils/ctx';
-function groupAssignmentsByDate(assignments: Assignment[]) {
 
+function groupAssignmentsByDate(assignments: Assignment[]) {
   const grouped: { title: string; data: Assignment[] }[] = [];
 
   // Tạo bản đồ nhóm bài tập theo ngày
   const map = new Map<string, Assignment[]>();
   assignments.forEach((assignment) => {
-    assignment.deadline = assignment.deadline + "Z"; // chờ backend fix là bỏ
+    assignment.deadline = assignment.deadline;
     const dateKey = new Date(assignment.deadline).toLocaleDateString('vi-VN', {
       timeZone: Intl.DateTimeFormat().resolvedOptions().timeZone,
     });
-
-    console.log(dateKey)
 
     if (!map.has(dateKey)) {
       map.set(dateKey, []);
     }
     map.get(dateKey)!.push(assignment);
-    console.log(assignment.deadline);
   });
 
   // Chuyển bản đồ thành mảng
   map.forEach((data, title) => {
     // **Sắp xếp các bài tập theo thời gian deadline**
-    data.sort((a, b) => new Date(a.deadline).getTime() - new Date(b.deadline).getTime());
+    data.sort((a, b) => new Date(b.deadline).getTime() - new Date(a.deadline).getTime());
     grouped.push({ title, data });
   });
 
-  // Sắp xếp các nhóm theo ngày gần nhất
+  // Sắp xếp các nhóm theo ngày trễ gần nhất
   grouped.sort((a, b) => {
     const [dayA, monthA, yearA] = a.title.split('/').map(Number); // Tách ngày/tháng/năm
     const [dayB, monthB, yearB] = b.title.split('/').map(Number);
 
     // So sánh đối tượng Date
-    return new Date(yearA, monthA - 1, dayA).getTime() - new Date(yearB, monthB - 1, dayB).getTime();
+    return new Date(yearB, monthB- 1, dayB).getTime() - new Date(yearA, monthA - 1, dayA).getTime();
   });
-  
+
   return grouped;
 }
 
-const UpcomingAssignments: React.FC = () => {
+const OverdueAssignments: React.FC = () => {
   const [sections, setSections] = useState<{ title: string; data: Assignment[] }[]>([]);
   const [loading, setLoading] = useState(true);
   const [refreshing, setRefreshing] = useState(false); // Trạng thái làm mới
@@ -53,7 +50,7 @@ const UpcomingAssignments: React.FC = () => {
   const loadAssignments = async () => {
     try {
       setLoading(true);
-      const data = await fetchAssignments('UPCOMING');
+      const data = await fetchAssignments('PASS_DUE');
       const groupedData = groupAssignmentsByDate(data);
       setSections(groupedData);
     } catch (err: any) {
@@ -67,7 +64,7 @@ const UpcomingAssignments: React.FC = () => {
   const handleRefresh = async () => {
     setRefreshing(true);
     try {
-      const data = await fetchAssignments('UPCOMING');
+      const data = await fetchAssignments('PASS_DUE');
       const groupedData = groupAssignmentsByDate(data);
       setSections(groupedData);
     } catch (err: any) {
@@ -122,9 +119,9 @@ const UpcomingAssignments: React.FC = () => {
         if (diffDays > 0) {
           comparisonText = `Còn ${diffDays} ngày`;
         } else if (diffDays === 0) {
-          comparisonText = `Hết hạn hôm nay`;
+          comparisonText = `Quá hạn hôm nay`;
         } else {
-          comparisonText = `Quá hạn ${Math.abs(diffDays)} ngày`;
+          comparisonText = `Đến hạn ${Math.abs(diffDays)} ngày trước`;
         }
       
         return (
@@ -138,6 +135,7 @@ const UpcomingAssignments: React.FC = () => {
       renderItem={({ item }) => (
         <AssignmentCard
           className={`${item.class_name}`}
+          classId={`${item.class_id}`}
           assignmentTitle={item.title}
           //dueDate={new Date(item.deadline).toLocaleDateString('vi-VN')}
           dueTime={new Date(item.deadline).toLocaleTimeString('vi-VN', {
@@ -145,10 +143,11 @@ const UpcomingAssignments: React.FC = () => {
             minute: '2-digit',
           })}
           isSubmitted={item.is_submitted}
+          isOverdue={true}
           onPress={() => {
             router.push({
-              pathname: '/student/classes/[classId]/assignments/[assignmentId]',
-              params: { classId: item.class_id, assignmentId: item.id, assignment: JSON.stringify(item)},
+              pathname: '/student/(tabs)/assignments/[assignmentId]',
+              params: { classId: item.class_id, assignmentId: item.id, assignment: JSON.stringify(item), overdue: "true"},
             });
           }}
         />
@@ -204,4 +203,4 @@ const styles = StyleSheet.create({
   },  
 });
 
-export default UpcomingAssignments;
+export default OverdueAssignments;
