+++ app/student/classes/[classId]/assignments/_layout.tsx
@@ -1,5 +1,37 @@
+import Header from '@/components/Header';
 import { Stack } from 'expo-router';
 
-export default function AssignmentLayout() {
-    return <Stack screenOptions={{ headerShown: false }}></Stack>;
+export default function ClassLayout() {
+    return (
+        <Stack>
+            <Stack.Screen
+                name="(tabs)"
+                options={{
+                    headerStyle: {
+                        backgroundColor: '#c21c1c',
+                    },
+                    headerTintColor: '#fff',
+                    headerTitleStyle: {
+                        fontWeight: 'bold',
+                    },
+                    headerTitleAlign: 'center',
+                    headerTitle: () => <Header title="Danh sách bài tập"></Header>,
+                }}
+            ></Stack.Screen>
+            <Stack.Screen
+                name="[assignmentId]"
+                options={{
+                    headerStyle: {
+                        backgroundColor: '#c21c1c',
+                    },
+                    headerTintColor: '#fff',
+                    headerTitleStyle: {
+                        fontWeight: 'bold',
+                    },
+                    headerTitleAlign: 'center',
+                    headerTitle: () => <Header title="Chi tiết bài tập"></Header>,
+                }}
+            ></Stack.Screen>
+        </Stack>
+    );
 }
